Moved run_manager to its own file.
